+++ frontend/src/metabase/collections/containers/CreateCollectionForm/index.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/no-default-export
+export { default } from "./CreateCollectionForm";
