+++ frontend/src/metabase/collections/containers/CreateCollectionForm/tests/CreateCollectionForm-Common.unit.spec.tsx
@@ -1,0 +1,37 @@
+import userEvent from "@testing-library/user-event";
+import { screen } from "__support__/ui";
+import { setup } from "./common";
+
+describe("CreateCollectionForm", () => {
+  it("displays correct blank state", () => {
+    setup();
+
+    expect(screen.getByLabelText("Name")).toBeInTheDocument();
+    expect(screen.getByLabelText("Name")).toHaveValue("");
+
+    expect(screen.getByLabelText("Description")).toBeInTheDocument();
+    expect(screen.getByLabelText("Description")).toHaveValue("");
+
+    expect(screen.getByText(/Collection it's saved in/i)).toBeInTheDocument();
+    expect(screen.getByText("Our analytics")).toBeInTheDocument();
+
+    expect(screen.getByRole("button", { name: "Cancel" })).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Create" })).toBeInTheDocument();
+  });
+
+  it("can't submit if name is empty", () => {
+    setup();
+    expect(screen.getByRole("button", { name: "Create" })).toBeDisabled();
+  });
+
+  it("calls onCancel when cancel button is clicked", () => {
+    const { onCancel } = setup();
+    userEvent.click(screen.getByRole("button", { name: "Cancel" }));
+    expect(onCancel).toHaveBeenCalledTimes(1);
+  });
+
+  it("does not show authority level controls", () => {
+    setup();
+    expect(screen.queryByLabelText("Collection type")).not.toBeInTheDocument();
+  });
+});
