+++ frontend/src/metabase/collections/containers/CreateCollectionForm/tests/common.tsx
@@ -1,0 +1,49 @@
+import { TokenFeatures, User } from "metabase-types/api";
+import {
+  createMockCollection,
+  createMockTokenFeatures,
+  createMockUser,
+} from "metabase-types/api/mocks";
+import { createMockState } from "metabase-types/store/mocks";
+import { setupEnterprisePlugins } from "__support__/enterprise";
+import { setupCollectionsEndpoints } from "__support__/server-mocks";
+import { mockSettings } from "__support__/settings";
+import { createMockEntitiesState } from "__support__/store";
+import { renderWithProviders } from "__support__/ui";
+import CreateCollectionForm from "../CreateCollectionForm";
+
+const ROOT_COLLECTION = createMockCollection({
+  id: "root",
+  name: "Our analytics",
+  can_write: true,
+});
+
+type SetupOpts = {
+  user?: User;
+  tokenFeatures?: TokenFeatures;
+};
+
+export function setup({
+  user = createMockUser({ is_superuser: true }),
+  tokenFeatures = createMockTokenFeatures(),
+}: SetupOpts = {}) {
+  const settings = mockSettings({ "token-features": tokenFeatures });
+  const onCancel = jest.fn();
+
+  setupEnterprisePlugins();
+  setupCollectionsEndpoints({
+    collections: [],
+    rootCollection: ROOT_COLLECTION,
+  });
+  renderWithProviders(<CreateCollectionForm onCancel={onCancel} />, {
+    storeInitialState: createMockState({
+      currentUser: user,
+      settings,
+      entities: createMockEntitiesState({
+        collections: [ROOT_COLLECTION],
+      }),
+    }),
+  });
+
+  return { onCancel };
+}
